+++ vm/actor/src/builtin/miner/mod.rs
@@ -47,7 +47,7 @@ use ipld_amt::Amt;
 use ipld_blockstore::BlockStore;
 use message::Message;
 use num_bigint::bigint_ser::{BigIntDe, BigIntSer};
-use num_bigint::{BigInt};
+use num_bigint::BigInt;
 use num_derive::FromPrimitive;
 use num_traits::{FromPrimitive, Zero};
 use runtime::{ActorCode, Runtime};
@@ -549,7 +549,7 @@ impl Actor {
             Ok(newly_vested_amount)
         })??;
 
-        notify_pledge_change(rt, &BigInt::from(newly_vested_amount).neg())?;
+        notify_pledge_change(rt, &newly_vested_amount.neg())?;
         let mut bf = BitField::new();
         bf.set(params.sector_number as usize);
 
@@ -770,10 +770,7 @@ impl Actor {
                 Ok(newly_vested_fund)
             })??;
 
-            notify_pledge_change(
-                rt,
-                &(BigInt::from(initial_pledge) - BigInt::from(vested_amount)),
-            )?;
+            notify_pledge_change(rt, &(initial_pledge - vested_amount))?;
         }
         Ok(())
     }
@@ -1197,7 +1194,7 @@ impl Actor {
                 })?;
             Ok(newly_vested_amount)
         })??;
-        let delta = BigInt::from(amount) - BigInt::from(vested_amount);
+        let delta = amount - vested_amount;
         notify_pledge_change(rt, &delta)?;
         Ok(())
     }
@@ -1307,7 +1304,7 @@ impl Actor {
             &amount_withdrawn,
         )?;
 
-        notify_pledge_change(rt, &BigInt::from(vested_amount).neg())?;
+        notify_pledge_change(rt, &vested_amount.neg())?;
 
         st.assert_balance_invariants(&rt.current_balance()?);
         Ok(())
@@ -1353,7 +1350,7 @@ where
             Ok(newly_vested_fund)
         })??;
 
-    notify_pledge_change(rt, &BigInt::from(vested_amount).neg())?;
+    notify_pledge_change(rt, &vested_amount.neg())?;
 
     // Note: because the cron actor is not invoked on epochs with empty tipsets, the current epoch is not necessarily
     // exactly the final epoch of the period; it may be slightly later (i.e. in the subsequent period).
@@ -2357,7 +2354,7 @@ where
     RT: Runtime<BS>,
 {
     burn_funds(rt, amount)?;
-    notify_pledge_change(rt, &BigInt::from(amount.clone()).neg())
+    notify_pledge_change(rt, &amount.clone().neg())
 }
 
 fn burn_funds<BS, RT>(rt: &mut RT, amount: &TokenAmount) -> Result<(), ActorError>
