+++ vm/actor/src/builtin/paych/mod.rs
@@ -216,16 +216,13 @@ impl Actor {
             st.lane_states[idx].redeemed = sv.amount;
 
             // 4. check operation validity
-            let new_send_balance = BigInt::from(st.to_send.clone()) + balance_delta;
+            let new_send_balance = st.to_send.clone() + balance_delta;
 
-            if new_send_balance < TokenAmount::from(0u8){
-                return
-                Err(
-                    ActorError::new(
-                        ExitCode::ErrIllegalState,
-                        "voucher would leave channel balance negative".to_owned(),
-                    )
-                );
+            if new_send_balance < TokenAmount::from(0u8) {
+                return Err(ActorError::new(
+                    ExitCode::ErrIllegalState,
+                    "voucher would leave channel balance negative".to_owned(),
+                ));
             }
 
             if new_send_balance > curr_bal {
